Add zookeeper types and support tasks
Due to some really unfortunate design choices in Zookeeper, our TOSCA types
need to include special quorum virtual node that simply groups the machines on
which zookeeper instances will run.

Quorum mainly serves as a barrier that makes sure all servers are up before we
start installing Zookeeper instances, because this is the only reliable way of
obtaining complete set of addresses of machines that needs to be placed in
Zookeeper configuration file.