Make version handling manual process again
It turned out that having version number handling rely solely on scm is
a bad idea, since such approaches generally fail when package is used in
isolation (with no scm information present). Installation from Github
archives are broken in this situation.

In order to centralize version number handling for package and all
associated tools, we moved it inside package.

Tool that generates release plugin descriptions was already modified to
work with changes, as were developer docs.

Change-Id: I23ecfa311a8ab0dd9f8f3e3ded0b251cb4834c5a