Add support for list merging into generator
Since some of the resources that are now merged are lists (most
prominent example of this is `imports` section of blueprint), we also
need to support merging those.

Note that current implementation discards duplicates from merged list.
This is done by design, since there is no need to have duplicated
stuff in lists as of now. But this is something one should keep in
mind if such functionality is needed in the future.

Change-Id: Ia7dfd550e22ef13826ca32f162190c288b91fedc