Add additional firewall rule for node manager
Node manager usually spawns container manager that listens to some
random ephemeral port. Because this is really not optimal in
environments with strict firewall rules, we assume that setup is done
in such a way that managers listen on selected port.

Note: Port must be kept in sync between cookbooks and TOSCA library.

Change-Id: Iea59c8d25949afd034f4cd4d425b77e5134cb91b