Fix Apache Cassandra type
When testing new functionality of Chef plugin for Cloudify, we noticed
that things were failing apart in suspicious ways. A quick look at the
Cassandra TOSCA revealed that there are quite a few things out of
place.

This commit fixes this by properly separating creation and
configuration steps. Additionally, we made sure that vm that has
Cassandra installed is properly configured as far as FQDNs goes.

Change-Id: I4bbeb40620f3fc7ad0e6e0741e59e2205dd9748d