Force apt update before modifying VM in any way
As we learned a hard way, there is a non-zero possibility of end user
trying to use Ubuntu image that has its unzip apt-cache entry invalid.
And when this happens, dice_common recipe will fail to converge since
apt will not be able to install unzip.

In order to always have a consistent base to install software, we now
force apt cache update as a first thing when installing any kind of
software.

This might seem a bit heavy handed approach, but since just about any
non-trivial operation will require installing something using package
manager, this change is not as drastic as it might seem at first.

Change-Id: I8067b83c2c4d0415476d322da4ff4a60931a687a