Remove Storm submitter --config parameter
This configuration parameter is not as useful as we wished it would
be, since none of the options specified in this file make it to the
submitter.

In order to keep the same level of configurability for topology, this
commit replaces configuration file with series of command-line
arguments to submitter. As an additional benefit, submitter now takes
default worker configuration into account when topology is executed.

Change-Id: Ic42679b6aab507da5c8428f3f73391135294ff77