Add CentOS hosts support to library
Before this change, there was no way of writing blueprint that would
use both Ubuntu and CentOS hosts of the same size, because we tied
instance size and operating system in one indivisible unit.

This commit changes that by decoupling instance size from OS that is
running inside it. Currently, we support Ubuntu and CentOS operating
systems, but adding support for new OS is almost trivial now.

Change-Id: I028971c6bf2227fe6442069537edd2f41db8760e