Generalize resource download task
Previous incarnation of resource downloading tasks was only capable of
downloading a single resource, which was limiting in all sorts of
ways. For example, by only being capable of handling single resource,
there was a hard limit on five resources max per node, since we only
have five standard life cycle events that we can hook onto during
installation.

New resource download task can handle arbitrary number of resources
that are passed in as an iterable of two element lists/tuples. First
element is key under which downloaded resource location will be
recorded in runtime properties and second element is the source of the
resource (URL or relative pat for resources that are bundled with
blueprint).

In order to support blueprints that can handle missing resource,
resource sources that are null or empty string are skipped. Res of the
installation can determine if the resource has been skipped by testing
for presence of key in runtime properties of the node.

Change-Id: I7659f2e69a17d4f3af6dd9d9f1c601ca14f79b13