Make it possible to set node roles
Blueprint author can now set roles that should be registered for this
node on DMon. By default, role list is empty.

We must make sure when specifying roles for the node that DMon
supports them. Specifying role that is unknown to DMon can lead to
hard to debug behavior, since there is no way of validating the roles
or determining what will happen if the invalid role is sent to DMon.

Change-Id: I561cdbdba7f4e17f0ce6b8c9c24f2870956b4f80