Add support for configuration file
Authentication and other static aspects of plugin can now be
configured using /etc/dice/dice.yaml configuration file. This reduces
the clutter in blueprint file, since user does not need to list them
for each node that needs communication with underlying platform's
services.

It is still possible to specify some or all of the settings in
blueprint file and in this case they will override the values from
configuration file. This can be used to customize deployment behavior
on a blueprint-by-blueprint basis.

Change-Id: I4d820854f37e2e0f9bb0cac4bb06853594917c57