Update plugin definition generator
Since our final goal is to host plugin package and definition as
artifacts on Github release page, we need some tool that merges all of
our library file into single include.

Because we flattened the library layout, getting all library files is
now as simple as listing folder contents.

Additionally, we added a safety check in merge routine that will
inform us if we managed to produce key conflict between our various
library files. Probability of this happening is fairly low, since we
mostly work with types in distinct namespaces, but better to be safe
than sorry.

Change-Id: I0dfeff11e9d94acfe5e2306175c06de83c5d04cb