Update library definitions of firewall rules
Updated library renamed some of the fields in firewall definition and
made others mandatory. This commit gets library types back in sync
with the base firewall definition.

Note that end users that are using library-provided types are not
affected by this change, since this is an implementation detail change
as far as they are concerned.

Change-Id: Iee03ba6f1b5af8cd3511b0923185f419d75c0153