Add support for Amazon AWS
This commit adds AWS to list of supported platforms. All of the
functionality that is present on OpenStack and FCO is also present in
AWS part, which makes it possible to deploy the same blueprint on all
platforms and even more importantly, prepare a blueprint that deploy
applications in multiple clouds.

Change-Id: I51d9bb55a40c645266756b11ab137682117d0589