Reorganize documentation to make it extendable
Existing documentation layout used some hard-coded tocs that were
quite annoying to edit each time we add/remove part of the
documentation. New layout make use of the glob capability of the
Sphinx and delegates the job of collecting documents to the tool.

Ability to extend the documentation without manual intervention is
also one of the prerequisites for automatically generating API
documentation for various TOSCA library components.

Change-Id: Idbacd3d934232328c0ba48d98427035342064d88