Serialize server creation process on FCO
Since FCO is having troubles when multiple server spawn requests are
sent to it in rapid succession, we serialized those requests using
file lock. Now new server creation requests is sent only when the
previous one has been processed.

Note that we only serialized server creation requests, since our tests
indicate that this is the only API call that is sensitive to breakage.
Waiting for server to fully deploy is still done in parallel.

Change-Id: I5c59b3a61de884f540cd162115b74c793e73eff8