Remove docker from lite plugin
Since docker server requires Chef components that are not part of the
lite version of the library, docker can only be used when full version
of the library is utilised by blueprint.

Due to the error in the last update of plugin description generator,
this error failed to manifest itself when testing lite version of the
library because docker definitions were only included in full
versions. This fix makes this behavior intentional.

Change-Id: If607b273f76fc64a2aa9f75b0c2baa307736075b