Add support for docker command templates
In previous implementation of container creation, command to execute
was fixed at deployment creation time, which severely limited
usefulness of containers, since it was impossible to reference any
other node in the blueprint.

To remedy this situation, command is now a template that gets expanded
at container creation time with requested container attributes. The
attributes can be set in preconfigure lifecycle operation.

Change-Id: I692124bc9f66ce4b00e263bb308f25fa2fac2773