Add zookeeper source inputs to the library
At the moment, we need to wait for new releases of Chef cookbooks and
TOSCA library in order to update a component. In case of critical
bugs, this may introduce quite a sizeable delay, which is bad.

Exposing zookeeper sources through inputs make it possible to deploy
updated components that have compatible interfaces by simply adjusting
inputs on the blueprint.

Change-Id: Ia3216f64ae36505d239679fb2b004d9eb19f94aa