Add operation inputs for server and image creation
Directly accessing node properties from lifecycle operations has some
major drawbacks when it comes to using intrinsic functions, since
those functions are only evaluated once when the deployment is
instantiated on Cloudify Manager.

In order to bypass this limitation, we can pass property values as
inputs to the operation. And since intrinsic functions are evaluated
before the operation is executed, implementation actually receive the
result of intrinsic function instead of their textual representation.

This commit only modifies access to the properties that are likely to
benefit from being able to utilize get_attribute and concat intrinsic
functions. In the perfect world, operation implementations would not
rely on Cloudify's context at all, but this is currently infeasible to
achieve, since we still need context in order to be able to retrieve
resources that are bundled in blueprint.

Change-Id: Ia2faa270bf4b5e8276c648ce9a930f41b31294d5