Pin openstacksdk version
Latest (0.10.0) release of openstacksdk requires netifaces package
that needs gcc and python development package to build, which is not
installed on the Cloudify Manager by default. Even worse, requiring
that gcc is installed on each host that will use our plugin is clearly
an overkill.

This commit introduces a temporary fix: we pined down the
openstacksdk package version to the latest one that manages to get by
without the netifaces package. This will resolve the problems for now
with no side-effects.

Long-term solution to this problem is to migrate to Cloudify 4.x that
solves the issues with gcc not being present on manager. As for the
not having gcc on application hosts, properly solving this would
require splitting the plugin into more manageable pieces, but this can
probably only be done if/when the orchestrator will be able to select
proper node type from abstract template.

Change-Id: Ib393d960e81f7a6646deff95ec7438b0153af2ec