Allows firewall ports to be specified as strings
In order to make add a bit more convenience to the blueprint authoring
process, we now support ports in firewall rules being specified as
strings on all platforms.

Note that this already worked for Amazon and OpenStack whose API that
accept strings in such places. In order to keep things consistent, we
added explicit conversion to all platforms.

Change-Id: I94e63f1313df96cbb2f0db370d85c828e3d5c0f9